+++ xenon/includes/xenon/utility/vec2.hpp
@@ -4,12 +4,12 @@
 
 class Vec2 {
 public:
-	double x, y;
+	float x, y;
 
 	Vec2() : x(-99), y(-99) {}
-	Vec2(double x, double y) : x(x), y(y) {}
+	Vec2(float x, float y) : x(x), y(y) {}
 
-	double Distance(Vec2 v) {
-		return sqrt(pow(v.x - x, 2) + pow(v.y - y, 2));
+	float Distance(Vec2 v) {
+		return static_cast<float>(sqrt(pow(v.x - x, 2) + pow(v.y - y, 2)));
 	}
 };
