+++ xenon/src/core/system.cpp
@@ -11,7 +11,7 @@ Vec2 System::GetScreenResolution() {
 	int width, height;
 	GetDesktopResolution(width, height);
 
-	Vec2 resolution{ static_cast<double>(width), static_cast<double>(height) };
+	Vec2 resolution{ static_cast<float>(width), static_cast<float>(height) };
 
 	//spdlog::debug("Screen resolution: {}x{}", resolution.x, resolution.y);
 
@@ -22,7 +22,7 @@ Vec2 System::GetScreenCenter() {
 	int width, height;
 	GetDesktopResolution(width, height);
 
-	Vec2 center{ static_cast<double>(width / 2), static_cast<double>(height / 2) };
+	Vec2 center{ static_cast<float>(width / 2), static_cast<float>(height / 2) };
 
 	//spdlog::debug("Screen center: {}, {}", center.x, center.y);
 
@@ -34,7 +34,7 @@ Vec2 System::GetMousePos() {
 	POINT p;
 	GetCursorPos(&p);
 
-	return Vec2{ static_cast<double>(p.x), static_cast<double>(p.y) };
+	return Vec2{ static_cast<float>(p.x), static_cast<float>(p.y) };
 
 }
